refactor: split DTOs out into Core
* keeps the DTOs visible where needed and easy to maintain
