+++ src/InMeal/InMeal.Core/DTOs/AddRecipeIngredientDto.cs
@@ -1,0 +1,5 @@
+using InMeal.Core.Entities;
+
+namespace InMeal.Core.DTOs;
+
+public record AddRecipeIngredientDto(Guid IngredientId, Quantity Quantity);
