+++ src/InMeal/InMeal.Core/Entities/RecipeIngredient.cs
@@ -14,5 +14,3 @@ public class RecipeIngredient
     // A JSON encoded field (non-nullable)
     public Quantity Quantity { get; set; }
 }
-
-public record AddRecipeIngredientDto(Guid IngredientId, Quantity Quantity);
